Moving tests classes to the right folders
